fix copy/paste/cut if an editable is focused - #57274
Copy, Paste and Cut work different when activated via key, menu or via
the popup menu associated with the currently edited cell. The reason
is that the key and main menu based clipboard actions are always working
on a complete single cell or a selection of cells. When a single cell is focused,
i.e. you edit the cell contents, then this is confusing because the normal
GUI behavior is that you work on the current text in the editable. That means
you can cut/copy/paste just some characters.

For Copy and Paste the behavior is changed when a cell is currently focused.
Then the copy and paste operation work on the current text selection in the
editable.
The Cut is currently done solely in pspp on the store without calling anything
in ssw. Therefore I added a new API method

  ssw_sheet_try_cut

which checks if a cell is currently focused. If yes, the cut operation is
done on the editable text and TRUE is returned. This can be used in pspp.
When ssw_sheet_try_cut is successfull, then the clipboard cut is already
done.

Closes: https://savannah.gnu.org/bugs/?57274